+++ extensions/amazon-aws/src/cloudformation.tsx
@@ -1,53 +1,24 @@
-import { getPreferenceValues, ActionPanel, List, Detail, Action } from "@raycast/api";
-import { useState, useEffect } from "react";
+import { ActionPanel, List, Detail, Action } from "@raycast/api";
 import * as AWS from "aws-sdk";
-import { Preferences } from "./types";
 import { StackSummary } from "aws-sdk/clients/cloudformation";
 import setupAws from "./util/setupAws";
+import { useCachedPromise } from "@raycast/utils";
 
-setupAws();
+const preferences = setupAws();
+const cloudformation = new AWS.CloudFormation({ apiVersion: "2016-11-15" });
 
 export default function ListStacks() {
-  const preferences: Preferences = getPreferenceValues();
-  const cloudformation = new AWS.CloudFormation({ apiVersion: "2016-11-15" });
+  const { data: stacks, error, isLoading } = useCachedPromise(fetchStacks);
 
-  const [state, setState] = useState<{
-    stacks: StackSummary[];
-    loaded: boolean;
-    hasError: boolean;
-  }>({
-    stacks: [],
-    loaded: false,
-    hasError: false,
-  });
-
-  useEffect(() => {
-    if (!preferences.region) return;
-    async function fetchStacks(token?: string, stacks?: StackSummary[]): Promise<StackSummary[]> {
-      const { NextToken, StackSummaries } = await cloudformation.listStacks({ NextToken: token }).promise();
-      const combinedStacks = [...(stacks || []), ...(StackSummaries || [])];
-
-      if (NextToken) {
-        return fetchStacks(NextToken, combinedStacks);
-      }
-
-      return combinedStacks.filter((stack) => stack.StackStatus !== "DELETE_COMPLETE");
-    }
-
-    fetchStacks()
-      .then((stacks) => setState({ hasError: false, loaded: true, stacks }))
-      .catch(() => setState({ hasError: true, loaded: false, stacks: [] }));
-  }, []);
-
-  if (state.hasError) {
+  if (error) {
     return (
       <Detail markdown="No valid [configuration and credential file](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html) found in your machine." />
     );
   }
 
   return (
-    <List isLoading={!state.loaded} searchBarPlaceholder="Filter stacks by name...">
-      {state.stacks.map((s) => (
+    <List isLoading={isLoading} searchBarPlaceholder="Filter stacks by name...">
+      {stacks?.map((s) => (
         <CloudFormationStack key={s.StackId} stack={s} />
       ))}
     </List>
@@ -55,8 +26,6 @@ export default function ListStacks() {
 }
 
 function CloudFormationStack({ stack }: { stack: StackSummary }) {
-  const preferences: Preferences = getPreferenceValues();
-
   return (
     <List.Item
       id={stack.StackName}
@@ -85,4 +54,15 @@ function CloudFormationStack({ stack }: { stack: StackSummary }) {
       ]}
     />
   );
+}
+
+async function fetchStacks(token?: string, stacks?: StackSummary[]): Promise<StackSummary[]> {
+  const { NextToken, StackSummaries } = await cloudformation.listStacks({ NextToken: token }).promise();
+  const combinedStacks = [...(stacks || []), ...(StackSummaries || [])];
+
+  if (NextToken) {
+    return fetchStacks(NextToken, combinedStacks);
+  }
+
+  return combinedStacks.filter((stack) => stack.StackStatus !== "DELETE_COMPLETE");
 }