+++ extensions/amazon-aws/src/codepipelines.tsx
@@ -1,86 +1,24 @@
-import { getPreferenceValues, ActionPanel, List, Detail, Action, Icon } from "@raycast/api";
-import { useState, useEffect } from "react";
+import { ActionPanel, List, Detail, Action, Icon } from "@raycast/api";
 import * as AWS from "aws-sdk";
-import { Preferences } from "./types";
 import setupAws from "./util/setupAws";
+import { useCachedPromise } from "@raycast/utils";
+import { PipelineSummary } from "aws-sdk/clients/codepipeline";
 
-setupAws();
-
-const getExecutionState = (client: AWS.CodePipeline, pipelineName: string) =>
-  new Promise<AWS.CodePipeline.PipelineExecutionSummary | null>((resolve, reject) => {
-    return client.listPipelineExecutions(
-      {
-        pipelineName: pipelineName,
-      },
-      (err, data) => {
-        if (err) {
-          reject(err);
-        } else {
-          if (data.pipelineExecutionSummaries?.[0]) {
-            resolve(data.pipelineExecutionSummaries[0]);
-          } else {
-            resolve(null);
-          }
-        }
-      }
-    );
-  });
-
-type PipelineSummary = AWS.CodePipeline.PipelineSummary & {
-  execution?: AWS.CodePipeline.PipelineExecutionSummary | null;
-};
+const preferences = setupAws();
+const pipeline = new AWS.CodePipeline({ apiVersion: "2016-11-15" });
 
 export default function DescribeInstances() {
-  const preferences: Preferences = getPreferenceValues();
-  const pipeline = new AWS.CodePipeline({ apiVersion: "2016-11-15" });
+  const { data: pipelines, error, isLoading } = useCachedPromise(fetchPipelines);
 
-  const [state, setState] = useState<{
-    pipelines: PipelineSummary[];
-    loaded: boolean;
-    hasError: boolean;
-  }>({
-    pipelines: [],
-    loaded: false,
-    hasError: false,
-  });
-
-  useEffect(() => {
-    if (!preferences.region) return;
-    async function fetch(token?: string, accPipelines?: PipelineSummary[]): Promise<PipelineSummary[]> {
-      const { nextToken, pipelines } = await pipeline.listPipelines({ nextToken: token }).promise();
-      const combinedPipelines = [...(accPipelines || []), ...(pipelines || [])];
-
-      if (nextToken) {
-        return fetch(nextToken, combinedPipelines);
-      }
-
-      return combinedPipelines.filter((p) => !!p.name);
-    }
-
-    fetch()
-      .then((pipelines) => {
-        setState({ hasError: false, loaded: true, pipelines });
-
-        pipelines.forEach(async (p) => {
-          const execution = await getExecutionState(pipeline, p.name as string);
-          setState((state) => ({
-            ...state,
-            pipelines: state.pipelines.map((el) => (el.name === p.name ? { ...p, execution } : el)),
-          }));
-        });
-      })
-      .catch(() => setState({ hasError: true, loaded: false, pipelines: [] }));
-  }, []);
-
-  if (state.hasError) {
+  if (error) {
     return (
       <Detail markdown="No valid [configuration and credential file](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html) found in your machine." />
     );
   }
 
   return (
-    <List isLoading={!state.loaded} searchBarPlaceholder="Filter codepipelines by name...">
-      {state.pipelines.map((i) => (
+    <List isLoading={isLoading} searchBarPlaceholder="Filter codepipelines by name...">
+      {pipelines?.map((i) => (
         <CodePipelineListItem key={i.name} pipeline={i} />
       ))}
     </List>
@@ -88,9 +26,9 @@ export default function DescribeInstances() {
 }
 
 function CodePipelineListItem({ pipeline }: { pipeline: PipelineSummary }) {
-  const preferences: Preferences = getPreferenceValues();
+  const { data: execution } = useCachedPromise(fetchExecutionState, [pipeline.name]);
 
-  const status = pipeline?.execution?.status || "Idle";
+  const status = execution?.status || "Idle";
 
   return (
     <List.Item
@@ -128,3 +66,22 @@ const iconMap: { [key: string]: Icon } = {
   Succeeded: Icon.CircleProgress100,
   Stopped: Icon.CircleFilled,
 };
+
+async function fetchPipelines(token?: string, accPipelines?: PipelineSummary[]): Promise<PipelineSummary[]> {
+  const { nextToken, pipelines } = await pipeline.listPipelines({ nextToken: token }).promise();
+  const combinedPipelines = [...(accPipelines || []), ...(pipelines || [])];
+
+  if (nextToken) {
+    return fetchPipelines(nextToken, combinedPipelines);
+  }
+
+  return combinedPipelines.filter((p) => !!p.name);
+}
+
+async function fetchExecutionState(pipelineName?: string) {
+  if (!pipelineName) {
+    return;
+  }
+  const { pipelineExecutionSummaries } = await pipeline.listPipelineExecutions({ pipelineName }).promise();
+  return pipelineExecutionSummaries?.[0];
+}