+++ extensions/linear/src/unread-notifications.tsx
@@ -1,4 +1,4 @@
-import { getApplications, MenuBarExtra, open, openCommandPreferences } from "@raycast/api";
+import { getApplications, MenuBarExtra, open, openCommandPreferences, launchCommand, LaunchType } from "@raycast/api";
 import { NotificationResult } from "./api/getNotifications";
 import { updateNotification } from "./api/updateNotification";
 import View from "./components/View";
@@ -33,44 +33,60 @@ function UnreadNotifications() {
     await open(`https://linear.app/${urlKey}/inbox`, linearApp);
   }
 
+  const truncate = (text: string, maxLength: number) => {
+    const ellipsis = text.length > maxLength ? "…" : "";
+    return text.substring(0, maxLength).trim() + ellipsis;
+  };
+
   return (
     <MenuBarExtra
       title={getNotificationMenuBarTitle(unreadNotifications)}
       icon={getNotificationMenuBarIcon(unreadNotifications)}
       isLoading={isLoadingNotifications}
     >
-      <MenuBarExtra.Item
-        title="Open Inbox"
-        icon="linear.png"
-        shortcut={{ modifiers: ["cmd"], key: "o" }}
-        onAction={openInbox}
-      />
-      <MenuBarExtra.Separator />
-      <MenuBarExtra.Item
-        title={unreadNotifications.length !== 0 ? "Unread Notifications" : "No Unread Notifications"}
-      />
-      {unreadNotifications.map((notification) => {
-        const baseTitle = `${getNotificationTitle(notification)} by ${
-          notification.actor ? notification.actor.displayName : "Linear"
-        }`;
+      <MenuBarExtra.Section>
+        <MenuBarExtra.Item
+          title="Open Inbox"
+          icon="linear.png"
+          shortcut={{ modifiers: ["cmd"], key: "o" }}
+          onAction={openInbox}
+        />
+      </MenuBarExtra.Section>
 
-        return (
-          <MenuBarExtra.Item
-            key={notification.id}
-            icon={notification.actor ? getUserIcon(notification.actor) : "linear.png"}
-            title={notification.issue ? `${notification.issue.identifier}: ${baseTitle}` : baseTitle}
-            tooltip={notification.issue?.title}
-            onAction={() => openNotification(notification)}
-          />
-        );
-      })}
+      <MenuBarExtra.Section>
+        <MenuBarExtra.Item
+          title={unreadNotifications.length !== 0 ? "Unread Notifications" : "No Unread Notifications"}
+        />
 
-      <MenuBarExtra.Separator />
-      <MenuBarExtra.Item
-        title="Configure Command"
-        shortcut={{ modifiers: ["cmd"], key: "," }}
-        onAction={() => openCommandPreferences()}
-      />
+        {unreadNotifications.map((notification) => {
+          const baseTitle = `${getNotificationTitle(notification)} by ${
+            notification.actor ? notification.actor.displayName : "Linear"
+          }`;
+
+          return (
+            <MenuBarExtra.Item
+              key={notification.id}
+              icon={notification.actor ? getUserIcon(notification.actor) : "linear.png"}
+              title={baseTitle}
+              subtitle={notification.issue?.title ? truncate(notification.issue.title, 20) : ""}
+              tooltip={`${notification.issue?.identifier}: ${notification.issue?.title}`}
+              onAction={() => openNotification(notification)}
+            />
+          );
+        })}
+      </MenuBarExtra.Section>
+
+      <MenuBarExtra.Section>
+        <MenuBarExtra.Item
+          title="View All Notifications"
+          onAction={() => launchCommand({ name: "notifications", type: LaunchType.UserInitiated })}
+        />
+        <MenuBarExtra.Item
+          title="Configure Command"
+          shortcut={{ modifiers: ["cmd"], key: "," }}
+          onAction={() => openCommandPreferences()}
+        />
+      </MenuBarExtra.Section>
     </MenuBarExtra>
   );
 }