+++ extensions/my-daily-log/src/shared/formatDateToReadable.tsx
@@ -1,0 +1,9 @@
+export function formatDateToReadable(date: Date): string {
+  if (date.getDate() === new Date().getDate()) {
+    return "Today";
+  }
+  if (date.getDate() === new Date().getDate() - 1) {
+    return "Yesterday";
+  }
+  return date.toLocaleDateString();
+}