+++ extensions/snake/changelog.md
@@ -1,5 +1,7 @@
 # Snake Changelog
 
+## [Added screenshots] - 2022-11-17
+
 ## [Update] - 2022-05-03
 
 - Updated Raycast API to 1.33.0