+++ extensions/unsplash/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unsplash Changelog
 
+## [Added screenshots] - 2022-11-17
+
 ## [Update] - 2022-08-04
 
 - Updated Raycast API to 1.38.2